requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
Alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
streamlit==1.39.0
requests==2.32.3
reportlab==4.2.2
